Routing login page
Login page routes from the navbar and allows user to submit details.

This then runs a method from within the user class (models/user) which compares the username and password and if matching, instantiates a new User object.

Struggling to catch this User and work out how to tell the code where to send them on submit. Currently just sends them on a loop of logins.